set price precision/scale, add check examples

+++ src/main/resources/db/changelog/changelogs/table_goods_01.sql
@@ -2,9 +2,9 @@
 --changeset sbobrovich:create-table-goods
 CREATE TABLE public.goods
 (
-    id                BIGSERIAL   NOT NULL PRIMARY KEY UNIQUE,
-    description       VARCHAR(50) NOT NULL UNIQUE,
-    price             DECIMAL     NOT NULL CHECK ( price >= 0 ),
+    id                BIGSERIAL      NOT NULL PRIMARY KEY UNIQUE,
+    description       VARCHAR(50)    NOT NULL UNIQUE,
+    price             DECIMAL(20, 2) NOT NULL CHECK ( price >= 0 ),
     quantity_in_stock INTEGER CHECK ( quantity_in_stock >= 0 ),
-    wholesale_goods   BOOLEAN     NOT NULL default false
+    wholesale_goods   BOOLEAN        NOT NULL default false
 );
